+++ CHANGELOG.md
@@ -2,8 +2,8 @@ Unreleased
 ----------
 - Added `weighted_average` member to `data::v2::bars::Bar` type
 - Bumped `hyper` dependency to `1.0`
-- Bumped `websocket-util` dependency to `0.13`
-- Bumped `tokio-tungstenite` dependency to `0.23`
+- Bumped `websocket-util` dependency to `0.14`
+- Bumped `tokio-tungstenite` dependency to `0.26`
 
 
 0.29.0